Commit #1
This is a project for all new learners. it helps to grow in Django. in this project Crud functionality and the user authentication and authorization will me done by me and also add some bootstrap things like navbar , cards and used the Forms Functionality of Django it will automatic gives a structure of Form using Models in Django. refer this project it will help you to grow more in Django.